==> analysis/objects/03_tximport.Rmd <==
---
title:    "Comparative analysis of mouse strains for in vivo reprogramming"
subtitle: "Import RNA-seq counts"
author:   "Author: João Agostinho de Sousa"
date:     "Date: `r format(Sys.time(),'%d %B, %Y')`"
output:
  html_document:
    df_print: paged
    toc: TRUE
    toc_float: TRUE
    theme: default
    code_folding: hide
editor_options:
  chunk_output_type: console
---

# Load libraries
```{r}

library(tidyverse)
library(tximport)

```

# Load metadata
```{r}

load(file = "output/database/metadata.RData")

```

# Create tximport object
```{r}

# List all salmon counts
salmon_files <- 
  list.files("data/rnaseq/counts", recursive = T, pattern = "quant.sf", full.names = T) %>% 
  set_names(str_remove(., "data/rnaseq/counts/") %>% str_remove(., "/quant.sf"))

# Order to match with metadata and remove samples that were excluded before alignment
salmon_files <-
  salmon_files[as.character(metadata$`Sample ID`)] %>% 
  na.omit

# Transcript ID to gene ID table
tx2gene <- read_tsv("docs/genome/tx2gene.tsv", col_names = F)

# Create tximport object
tximport_object <- 
  tximport(files   = salmon_files, 
           type    = "salmon", 
           txIn    = T, 
           txOut   = F, 
           tx2gene = tx2gene)

```

# Save object
```{r}

save(tximport_object, file = "output/database/tximport_object.RData")

```

# Session information
```{r Session info, echo = FALSE, results = "markup"}

devtools::session_info()

```

This document was processed on: `r Sys.Date()`.
