==> analysis/objects/05_log2_norm_counts.Rmd <==
---
title:    "Comparative analysis of mouse strains for in vivo reprogramming"
subtitle: "Log2 normalized counts objects"
author:   "Author: João Agostinho de Sousa"
date:     "Date: `r format(Sys.time(),'%d %B, %Y')`"
output:
  html_document:
    df_print: paged
    toc: TRUE
    toc_float: TRUE
    theme: default
    code_folding: hide
editor_options:
  chunk_output_type: console
---

# Load libraries
```{r}

library(tidyverse)
library(SummarizedExperiment)

```

# Load objects
```{r}

# SummarizedExperiment
load(file = "output/database/SummarizedExperiment_object.RData")

```

# Create log2 normalized counts dataframes in a long table format
```{r}

# Log2 normalized counts in a long format
log2_norm_counts_long <- 
  assay(SummarizedExperiment_object, "log2_norm_counts") %>% 
    as.data.frame() %>% 
    rownames_to_column("gene_id") %>% 
    gather(`Sample ID`, log2_norm_counts, -gene_id) %>% 
    left_join(as.data.frame(rowData(SummarizedExperiment_object))) %>% 
    left_join(as.data.frame(colData(SummarizedExperiment_object)) %>% rownames_to_column("Sample ID"))

# Log2 normalized counts with mean and median values of the biological replicates
log2_norm_counts_long_summarised <- 
  log2_norm_counts_long %>% 
    group_by(gene_id, seqnames, start, end, width, strand, gene_name, gene_biotype, Tissue, Genotype, Color) %>% 
    summarise(median_log2_norm_counts = median(log2_norm_counts),
              mean_log2_norm_counts   = mean(log2_norm_counts),
              sd_log2_norm_counts     = sd(log2_norm_counts)) %>% 
    ungroup

```

# Save object
```{r}

log2_norm_counts_object <- 
list("log2_norm_counts_long"            = log2_norm_counts_long,
     "log2_norm_counts_long_summarised" = log2_norm_counts_long_summarised)

save(log2_norm_counts_object, file = "output/database/log2_norm_counts_object.RData")

```

# Session information
```{r Session info, echo = FALSE, results = "markup"}

devtools::session_info()

```

This document was processed on: `r Sys.Date()`.
